Invoke parent class constructor (SUPER::new)
Proximate reason is this warning in LWP::UserAgent 6.50.0(*)

   Use of uninitialized value in numeric ge (>=) at /usr/lib64/perl5/vendor_perl/5.20.1/LWP/UserAgent.pm line 285.
   Use of uninitialized value in concatenation (.) or string at /usr/lib64/perl5/vendor_perl/5.20.1/LWP/UserAgent.pm line 286.

This is because LWP::UserAgent::request() references $self->{max_redirect}
which is initialized in the constructor. We could just hardcode the same
initialization into WebService::RTMAgent::new(), but this is cleaner and
more futureproof.

(*) 6.50.0 on gentoo. There is no such version on CPAN. I don't get it
    either, but FWIW 6.13 on CPAN shows the same issue.

Signed-off-by: Ed Santiago <esm@cpan.org>